refactor ReturnBook to receive prop to populate backend information
This is the last commit out of 3 (starting in "use useEffect to fetch information from backend, create BookModel.ts") to retrieve information from backend. Backend needs to be running to test it. Change in theAllowedOrigins -> the CORS policy now properly allows your React frontend to communicate with the backend API.